Branch for fix version 6.3.4

+++ com.jaspersoft.studio/src/com/jaspersoft/studio/swt/widgets/NullableSpinner.java
@@ -1,6 +1,14 @@
 /*******************************************************************************
- * Copyright (C) 2010 - 2016. TIBCO Software Inc. 
- * All Rights Reserved. Confidential & Proprietary.
+ * Copyright (C) 2005 - 2014 TIBCO Software Inc. All rights reserved.
+ * http://www.jaspersoft.com.
+ * 
+ * Unless you have purchased  a commercial license agreement from Jaspersoft,
+ * the following license terms  apply:
+ * 
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
  ******************************************************************************/
 package com.jaspersoft.studio.swt.widgets;
 
@@ -423,7 +431,7 @@ public class NullableSpinner extends Composite {
 	 * @param value the new increment (must be greater than zero)
 	 * 
 	 */
-	public void setIncrement(float value) {
+	public void setIncrement(int value) {
 		this.checkWidget();
 		text.setIncrementStep(value);
 	}
